feat: renombrar storage.ts como image-fallbacks.ts, eliminar funciones residuales sin usar y refactorizar importaciones en componentes de la enciclopedia
